+++ src/monty/dev.py
@@ -7,9 +7,6 @@ from __future__ import annotations
 
 import functools
 import inspect
-import logging
-import os
-import subprocess
 import sys
 import warnings
 from dataclasses import is_dataclass
@@ -18,8 +15,6 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from typing import Callable, Optional, Type
-
-logger = logging.getLogger(__name__)
 
 
 def deprecated(
@@ -35,8 +30,7 @@ def deprecated(
         replacement (Callable | str): A replacement class or function.
         message (str): A warning message to be displayed.
         deadline (Optional[tuple[int, int, int]]): Optional deadline for removal
-            of the old function/class, in format (yyyy, MM, dd). A CI warning would
-            be raised after this date if is running in code owner' repo.
+            of the old function/class, in format (yyyy, MM, dd).
         category (Warning): Choose the category of the warning to issue. Defaults
             to FutureWarning. Another choice can be DeprecationWarning. Note that
             FutureWarning is meant for end users and is always shown unless silenced.
@@ -45,47 +39,8 @@ def deprecated(
             the choice accordingly.
 
     Returns:
-        Original function, but with a warning to use the updated function.
+        Original function/class, but with a warning to use the replacement.
     """
-
-    def raise_deadline_warning() -> None:
-        """Raise CI warning after removal deadline in code owner's repo."""
-
-        def _is_in_owner_repo() -> bool:
-            """Check if is running in code owner's repo.
-            Only generate reliable check when `git` is installed and remote name
-            is "origin".
-            """
-
-            try:
-                # Get current running repo
-                result = subprocess.run(
-                    ["git", "config", "--get", "remote.origin.url"],
-                    stdout=subprocess.PIPE,
-                )
-                owner_repo = (
-                    result.stdout.decode("utf-8")
-                    .strip()
-                    .lstrip("https://github.com/")  # HTTPS clone
-                    .lstrip("git@github.com:")  # SSH clone
-                    .rstrip(".git")  # SSH clone
-                )
-
-                return owner_repo == os.getenv("GITHUB_REPOSITORY")
-
-            except (subprocess.CalledProcessError, FileNotFoundError):
-                return False
-
-        # Only raise warning in code owner's repo CI
-        if (
-            _deadline is not None
-            and os.getenv("CI") is not None
-            and datetime.now() > _deadline
-            and _is_in_owner_repo()
-        ):
-            raise DeprecationWarning(
-                f"This function should have been removed on {_deadline:%Y-%m-%d}."
-            )
 
     def craft_message(
         old: Callable,
@@ -150,11 +105,8 @@ def deprecated(
 
         return cls
 
-    # Convert deadline to datetime type
-    _deadline = datetime(*deadline) if deadline is not None else None
-
-    # Raise CI warning after removal deadline
-    raise_deadline_warning()
+    # Convert deadline to `datetime` type
+    _deadline: datetime | None = datetime(*deadline) if deadline is not None else None
 
     def decorator(target: Callable) -> Callable:
         if inspect.isfunction(target):